// whirlpool/whirlpoolemea/cc-google-tag-manager/react/typings/ProductTypes.ts
export type ProductSpecification = {
  name: string;
  originalName: string;
  values: string[];
};

/** @deprecated Use `ProductSpecification` */
export type SpecificationGroupItem = ProductSpecification;

export type SpecificationGroup = {
  name: string;
  originalName: string;
  specifications: ProductSpecification[];
};

export type SecondaryRatingsAverages = {
  Id: string;
  AverageRating: number;
}

export type Product = {
  brand: string;
  brandId: string;
  cacheId: string;
  category: string;
  macroCategory: string;
  categories: string[];
  categoriesIds: string[];
  categoryId: string;
  categoryTree: Array<{ id: string; name: string; href: string }>;
  clusterHighlights: Array<{ id: string; name: string }>;
  description: string;
  itemMetadata: ItemMetadata;
  items: Item[];
  link: string;
  linkText: string;
  metaTagDescription: string;
  priceRange: {
    sellingPrice: { highPrice: number; lowPrice: number };
    listPrice: { highPrice: number; lowPrice: number };
  };
  productClusters: Array<{ id: string; name: string }>;
  productId: string;
  productName: string;
  productReference: string;
  properties: ProductSpecification[];
  skuSpecifications: SkuSpecification[];
  specificationGroups: SpecificationGroup[];
  titleTag: string;
  totalReviews: number;
  averageOverallRating: number;
  secondaryRatingsAverages: SecondaryRatingsAverages[];
};

export type Item = {
  complementName: string;
  ean: string;
  images: Array<{
    imageId: string;
    imageLabel: string;
    imageTag: string;
    imageUrl: string;
    imageText: string;
  }>;
  itemId: string;
  measurementUnit: string;
  name: string;
  nameComplete: string;
  referenceId: Array<{ Key: string; Value: string }>;
  sellers: Seller[];
  unitMultiplier: number;
  variations: Array<{ name: string; values: string[] }>;
  videos: Array<{ videoUrl: string }>;
};

export type SkuSpecification = {
  field: SkuSpecificationField;
  values: SkuSpecificationValues[];
};

export type SkuSpecificationField = {
  name: string;
};

export type SkuSpecificationValues = {
  name: string;
};

export type Seller = {
  sellerId: string;
  sellerName: string;
  addToCartLink: string;
  sellerDefault: boolean;
  commertialOffer: CommercialOffer;
};
// ...
export type ItemMetadata = {
  items: Array<{
    id: string;
    name: string;
    imageUrl: string;
    seller: string;
    assemblyOptions: Array<{
      id: string;
      name: string;
      required: boolean;
      inputValues: InputValue[];
      composition: Composition | null;
    }>;
  }>;
  priceTable: Array<{
    type: string;
    values: Array<{ id: string; assemblyId: string; price: number | null }>;
  }>;
};

export type InputValue = TextInputValue | BooleanInputValue | OptionsInputValue;
// ...
export type Composition = {
  minQuantity: number;
  maxQuantity: number;
  items: Array<{
    id: string;
    minQuantity: number;
    maxQuantity: number;
    priceTable: string;
    seller: string;
    initialQuantity: number;
  }>;
};